Move tests to the package directory
