+++ winter_messaging_transactional/tests/app_sample/consumer_correct/consumer_correct_event_handler.py
@@ -1,9 +1,9 @@
 from injector import inject
 from sqlalchemy.orm import Session
 
-from e2e.winter_messaging.app_sample.events import SampleEvent
+from winter_messaging_transactional.tests.app_sample.events import SampleEvent
 from winter.messaging import event_handler
-from e2e.winter_messaging.app_sample.dao import ConsumerDAO
+from winter_messaging_transactional.tests.app_sample.dao import ConsumerDAO
 
 
 class ConsumerCorrectEventHandler:
